+++ lib/generators/cms/component/language_switch/templates/app/cells/language_switch_cell.rb
@@ -1,7 +1,23 @@
 class LanguageSwitchCell < Cell::Rails
-  def show(page)
-    @languages = page.website.children()
-  	@current_language = page.homepage.locale
+  helper :cms
+
+  def show(homepages, current_homepage)
+    @homepages = homepages
+    @current_homepage = current_homepage
+
     render
   end
+
+  def entry(homepage)
+    @homepage = homepage
+
+    render
+  end
+
+  private
+
+  def current_homepage?(homepage)
+    homepage == @current_homepage
+  end
+  helper_method :current_homepage?
 end